Remove margin between file input field and link
The file input field is wrapped by a `div` with class `.form-item` and
thus gets a bottom margin. The `p` element containing the link to the
current file has a top margin. So there is a lot of space between input
field and link. This change sets these margins to `0`.